feat(metrics)!: Remove redundant transfer-rate prometheus metrics
This may be somewhat controvercial, so I'll do my best to
explain my rationale:

The existing prometheus metrics which describe per-session total
bytes uploaded and downloaded are substantially more accurate.

_Why?_

The transfer rate provided by qbit is a moving average over _some_
duration, and that duration isn't tied in any way to the
configurable polling rate that the prometheus metrics are scraped.
At best this can be derrived from existing metrics (as the value
is quite literally the derivative of the transferred bytes graph).
It's possible to read exclusively peaks or valleys, and fail to
get a sense of actual transfer rate.

Prometheus expects such data to be provided as counters (as qui
already exposes as per-session uploaded and downloaded bytes), so
that the timeseries database may calculate a rate. This guarantees
all bytes transferred are taken into account, effectively matching
the aformentioned sliding window to either the prometheus polling
interval, or a larger interval requested in prometheus queries.

Given this, the rate metric is something of a foot-gun, and we might
as well reduce the periodic API requests from prometheus scrapes
to slightly reduce server load.